% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actinoJSON.R
\name{convert_json_files_to_csv}
\alias{convert_json_files_to_csv}
\title{Convert to CSV all uBiome JSON files in a directory.}
\usage{
convert_json_files_to_csv(pattern = "[[:alnum:]].json", directory = getwd())
}
\arguments{
\item{pattern}{a regular expression specifying which files to look for (default is anything JSON)}

\item{directory}{the pathname of the directory to search (default is current working directory)}
}
\description{
Convenient way to automatically convert all JSON files in a directory to CSV.
}

