% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actinoCSV.R
\name{experiment_to_phyloseq}
\alias{experiment_to_phyloseq}
\title{convert a uBiome experiment dataframe and mapping file to a valid phyloseq object}
\usage{
experiment_to_phyloseq(experiment.df, mapfile, rank = "genus")
}
\arguments{
\item{experiment.df}{dataframe representation of a uBiome experiment. Rows are unique ssrs.}

\item{mapfile}{a dataframe with column 'ssr'}

\item{rank}{(not used currently -- just assumes everything is always just genus)}
}
\value{
a Phyloseq object
}
\description{
Make a valid Phyloseq object out of a uBiome dataframe and a map file
}
\examples{
simple.df <- data.frame(tax_name=c("taxa1"),"1234"=c(2352))
simple.map <- data.frame(ssr=c(1234))
}
