% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actinoJSON.R
\name{join_all_ubiome_files}
\alias{join_all_ubiome_files}
\title{returns a big dataframe joining all tax_names from a list of JSON files}
\usage{
join_all_ubiome_files(flist, tax_rank = "genus", count.normalized = FALSE)
}
\arguments{
\item{flist}{list of file pathnames, each pointing to a uBiome JSON file}

\item{tax_rank}{taxonomic rank to read. Default is "genus"}

\item{count.normalized}{use the count_norm field instead of raw read numbers (default is FALSE)}
}
\value{
data frame where first col is all taxa names, other cols are samples
}
\description{
returns a big dataframe joining all tax_names from a list of JSON files
}
