% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actinoJSON.R
\name{just_json_files_in}
\alias{just_json_files_in}
\title{Just JSON files in a directory}
\usage{
just_json_files_in(d)
}
\arguments{
\item{d}{pathname to a directory}
}
\description{
returns a vector of just the json files in a directory d
}
