% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actino_data.R
\docType{data}
\name{kombucha.csv}
\alias{kombucha.csv}
\title{microbiome data available to ub.phyloseq}
\format{An object of class \code{data.frame} with 277 rows and 6 columns.}
\usage{
data(kombucha.csv)
}
\description{
sample CSV file showing one kombucha sample
}
\keyword{microbiome}

