% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actino_data.R
\docType{data}
\name{kombucha}
\alias{kombucha}
\title{microbiome data available to ub.phyloseq}
\format{An object of class \code{phyloseq} of length 1.}
\usage{
data(kombucha)
}
\description{
Series of 19 samples during a kombucha experiment
}
\keyword{microbiome}

