% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actinoUtils.R
\name{matrix_at_rank}
\alias{matrix_at_rank}
\title{Matrix version of a dataframe at a specific rank}
\usage{
matrix_at_rank(df, rank = "genus")
}
\arguments{
\item{df}{a well-formed dataframe created by join_all_ubiome_files_full}

\item{rank}{tax_rank (default = "genus")}
}
\description{
turn a uBIome data frame into a matrix at a specific rank
}
