% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actinoJSON.R
\name{read_ubiome_json}
\alias{read_ubiome_json}
\title{Read a uBiome JSON file into a dataframe, returning only the taxa from a single rank.}
\usage{
read_ubiome_json(fname, rank = "genus", count.normalized = FALSE)
}
\arguments{
\item{fname}{file name as character vector}

\item{rank}{tax_rank as character vector}

\item{count.normalized}{use the count_norm field instead of raw read numbers (default is FALSE)}
}
\value{
data frame representation of the sample
}
\description{
Reads JSON into a dataframe.
}
