% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actinoCSV.R
\name{samples_as_column_dataframe}
\alias{samples_as_column_dataframe}
\title{Turn a dataframe into a format that Phyloseq likes}
\usage{
samples_as_column_dataframe(DF)
}
\arguments{
\item{DF}{data frame with columns: SSR, tax_name, count, (optional) tax_rank}
}
\description{
Convenience function to make a dataframe where sample names are in columns, not rows
}

