% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actinoJSON.R
\name{tax_abbrev}
\alias{tax_abbrev}
\title{returns the abbreviated character form of the character vector "tax_rank"}
\usage{
tax_abbrev(tax_rank, inverse = FALSE)
}
\arguments{
\item{tax_rank}{a character vector (e.g. "genus")}

\item{inverse}{if true, return the full tax_rank when given the abbreviated version}
}
\description{
Needed by functions that convert tax_names to QIIME (or other) format.
}

