% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actinoJSON.R
\name{tax_rank_of_full_taxa}
\alias{tax_rank_of_full_taxa}
\title{find tax_rank for a given full taxa}
\usage{
tax_rank_of_full_taxa(parse_taxon)
}
\arguments{
\item{parse_taxon}{string representation for qiime full taxonomy for this taxon}
}
\description{
opposite function for full_taxa(). Return a list [tax_name, tax_rank]
}

